fix(TDP-11262) : Fix the badly built 'isInvalid' and 'isEmpty' DSEL (#898)
expressions by the TQL to DSEL convertor. + Update of unit tests.

**What is the problem this Pull Request is trying to solve?**
isInvalid() and isEmpty() DSEL expressions are badly built by this
convertor. It cause a built failure of a DSEL expression from any TQL
expression that contains an isInvalid() and/or isEmpty() function with
at least one field with a semantic type.

**What is the chosen solution to this problem?**
Fix the Java method responsible for this.

**Link to the JIRA issue**
https://jira.talendforge.org/browse/TDP-11262